fix cross validation
