+++ python/dsbox/controller/config.py
@@ -147,8 +147,8 @@ class DsboxConfig:
 
     def _load_dsbox(self):
         self._load_logging()
-        # self.search_method = 'parallel'
-        self.search_method = 'serial'
+        self.search_method = 'parallel'
+        # self.search_method = 'serial'
 
     def _setup(self):
         self._define_create_output_dirs()
